Split IdentityVerificationReport.cs into multiple files
